refactor: validator 클래스 구현
입력한 자동차의 이름을 검증하는 기능을 별도의 class 로
분리하는 것이 역할을 더 명확히 한다 판단하여 분리함
