+++ src/main/java/me/yevgnenll/racing/UserInput.java
@@ -2,13 +2,13 @@ package me.yevgnenll.racing;
 
 import java.util.Scanner;
 
+import me.yevgnenll.racing.validator.NameValidator;
+
 public class UserInput {
 
 	private static final String USER_INPUT_CAR_NAME_MESSAGE =
 		"경주할 자동차의 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
 	private static final String USER_INPUT_COUNT_MESSAGE = "시도할 회수는 몇회인가요?";
-	private static final int NAME_MAX_LENGTH = 5;
-	private static final String NAME_SPLIT = ",";
 
 	private Scanner scanner = new Scanner(System.in);
 
@@ -22,21 +22,8 @@ public class UserInput {
 		print(USER_INPUT_CAR_NAME_MESSAGE);
 
 		String names = scanner.next();
-		checkValidationName(names);
+		NameValidator.validate(names);
 		return names;
-	}
-
-	protected void checkValidationName(String names) {
-		for (String name : names.split(NAME_SPLIT)) {
-			isLongerThan(name);
-		}
-	}
-
-	private boolean isLongerThan(String name) {
-		if (name.length() > NAME_MAX_LENGTH) {
-			throw new IllegalArgumentException("각 이름은 " + NAME_MAX_LENGTH + "자 이하여야 합니다.");
-		}
-		return false;
 	}
 
 	private int userInputCount() {
