+++ src/test/java/me/yevgnenll/racing/validator/NameValidatorTest.java
@@ -1,19 +1,19 @@
-package me.yevgnenll.racing;
+package me.yevgnenll.racing.validator;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class UserInputTest {
+import me.yevgnenll.racing.validator.NameValidator;
+
+public class NameValidatorTest {
 
 	@Test
 	@DisplayName(value = "입력된 자동차의 이름은 5글자 이내여야 한다")
 	void lessThan5() {
-		UserInput userInput = new UserInput();
-
 		assertThrows(IllegalArgumentException.class,
-			() -> userInput.checkValidationName("kaki,5longer,zeze"));
+			() -> NameValidator.validate("kaki,5length,zeze"));
 	}
 
 }
